+++ app/src/main/java/com/orego/corporation/orego/fragments/cameraFragment/CameraFragment.java
@@ -1,9 +1,12 @@
 package com.orego.corporation.orego.fragments.cameraFragment;
 
+import android.animation.AnimatorSet;
+import android.animation.ValueAnimator;
+import android.annotation.TargetApi;
 import android.app.Activity;
 import android.content.Intent;
 import android.graphics.Bitmap;
-import android.net.Uri;
+import android.os.Build;
 import android.os.Bundle;
 import android.provider.MediaStore;
 import android.support.annotation.NonNull;
@@ -12,28 +15,45 @@ import android.support.design.widget.BottomSheetBehavior;
 import android.support.design.widget.CoordinatorLayout;
 import android.support.v4.app.Fragment;
 import android.util.Log;
+import android.view.Gravity;
 import android.view.LayoutInflater;
+import android.view.SurfaceHolder;
+import android.view.SurfaceView;
 import android.view.View;
 import android.view.ViewGroup;
+import android.view.animation.LinearInterpolator;
+import android.widget.FrameLayout;
 import android.widget.ImageView;
 import android.widget.TextView;
 
 import com.orego.corporation.orego.R;
-import com.orego.corporation.orego.fragments.otherActivities.face3dActivity.model3D.view.ModelActivity;
-import com.orego.corporation.orego.views.cameraview.CameraView;
+import com.orego.corporation.orego.views.cameraview.CameraManager;
+import com.orego.corporation.orego.views.cameraview.CameraUtils;
 
 import java.io.File;
-import java.io.FileOutputStream;
-import java.io.IOException;
 import java.util.Objects;
 
-public class CameraFragment extends Fragment implements CameraView.CameraListener {
-    private static final String TAG = "CameraFragment";
-    private CameraView mCameraView;
-    private String mPath;
+import static android.view.View.GONE;
+import static android.view.View.VISIBLE;
 
+public class CameraFragment extends Fragment implements SurfaceHolder.Callback,
+        View.OnClickListener {
+    private static final String TAG = "CameraFragment";
+    private SurfaceView mSurfaceView;
+    private View mSwitchWrapper;
+    private ImageView mPictureView;
+    private CameraListener mCameraListener;
+    private Bitmap mPicture;
+    private boolean isSurfaceCreated;
+    private View captureRetryLayout;
+    private ImageView btnCapture;
+    private ImageView btnRetry;
+    private ImageView btnSwitchCamera;
+    private boolean isExpanded;
     private BottomSheetBehavior mBottomSheetBehavior;
     private TextView mTextViewState;
+    private ImageView btnSheetOpen;
+    private ImageView buttonCollapse;
 
     public static void startForResult(Activity activity, File path, int requestCode) {
         Intent intent = new Intent(activity, CameraFragment.class);
@@ -46,28 +66,19 @@ public class CameraFragment extends Fragment implements CameraView.CameraListene
     @Override
     public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
         CoordinatorLayout coordinatorLayout = (CoordinatorLayout) inflater.inflate(R.layout.camera_fragment, container, false);
-        mCameraView = coordinatorLayout.findViewById(R.id.camera_view);
-        mPath = Objects.requireNonNull(getActivity()).getIntent().getStringExtra(MediaStore.EXTRA_OUTPUT);
-        mCameraView.setCameraListener(this);
+//        mCameraView = coordinatorLayout.findViewById(R.id.camera_view);
+        //    private CameraView mCameraView;
+        String mPath = Objects.requireNonNull(getActivity()).getIntent().getStringExtra(MediaStore.EXTRA_OUTPUT);
+//        mCameraView.setCameraListener(this);
 
 
         View bottomSheet = coordinatorLayout.findViewById(R.id.bottom_sheet);
         mBottomSheetBehavior = BottomSheetBehavior.from(bottomSheet);
         mTextViewState = coordinatorLayout.findViewById(R.id.text_view_state);
-        ImageView btnSheetOpen = (ImageView) coordinatorLayout.findViewById(R.id.btn_sheet_open);
-        ImageView buttonCollapse = (ImageView) coordinatorLayout.findViewById(R.id.btn_sheet_close);
-        btnSheetOpen.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                mBottomSheetBehavior.setState(BottomSheetBehavior.STATE_EXPANDED);
-            }
-        });
-        buttonCollapse.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                mBottomSheetBehavior.setState(BottomSheetBehavior.STATE_COLLAPSED);
-            }
-        });
+        btnSheetOpen = (ImageView) coordinatorLayout.findViewById(R.id.btn_sheet_open);
+        buttonCollapse = (ImageView) coordinatorLayout.findViewById(R.id.btn_sheet_close);
+        btnSheetOpen.setOnClickListener(v -> mBottomSheetBehavior.setState(BottomSheetBehavior.STATE_EXPANDED));
+        buttonCollapse.setOnClickListener(v -> mBottomSheetBehavior.setState(BottomSheetBehavior.STATE_COLLAPSED));
         mBottomSheetBehavior.setBottomSheetCallback(new BottomSheetBehavior.BottomSheetCallback() {
             @Override
             public void onStateChanged(@NonNull View bottomSheet, int newState) {
@@ -97,60 +108,234 @@ public class CameraFragment extends Fragment implements CameraView.CameraListene
         });
 
 
+//        setClickable(true);
+        isExpanded = false;
+//        setBackgroundColor(Color.BLACK);
+        mCameraListener = new CameraListener() {
+            @Override
+            public void onCapture(Bitmap bitmap) {
+                Log.d(TAG, bitmap.toString());
+            }
+
+            @Override
+            public void onCameraError(Throwable th) {
+                Log.e(TAG, th.getMessage());
+            }
+        };
+        mSurfaceView = (SurfaceView) coordinatorLayout.findViewById(R.id.camera_surface);
+        mSurfaceView.setOnLongClickListener((v) -> {
+            onSwitchClick();
+            return true;
+        });
+        mSwitchWrapper = coordinatorLayout.findViewById(R.id.camera_switch_wrapper);
+        mPictureView = (ImageView) coordinatorLayout.findViewById(R.id.camera_picture_preview);
+
+        CameraManager.getInstance().init(getContext());
+
+        // fix `java.lang.RuntimeException: startPreview failed` on api 10
+        mSurfaceView.getHolder().addCallback(this);
+        mSwitchWrapper.setVisibility(CameraManager.getInstance().hasMultiCamera() ? VISIBLE : GONE);
+        mSwitchWrapper.setOnClickListener(this);
+
+        captureRetryLayout = coordinatorLayout.findViewById(R.id.camera_capture_retry_layout);
+
+        btnCapture = (ImageView) coordinatorLayout.findViewById(R.id.camera_capture);
+        btnRetry = (ImageView) coordinatorLayout.findViewById(R.id.camera_retry);
+        btnSwitchCamera = (ImageView) coordinatorLayout.findViewById(R.id.camera_close);
+
+        btnCapture.setOnClickListener(this);
+        btnRetry.setOnClickListener(this);
+        btnRetry.setEnabled(false);
+        btnSwitchCamera.setOnClickListener(this);
+
+
         return coordinatorLayout;
     }
 
 
     @Override
-    public void onCapture(Bitmap bitmap) {
-        File file = new File(mPath);
-        if (!file.getParentFile().exists()) {
-            file.getParentFile().mkdirs();
-        }
-
-        try {
-            FileOutputStream out = new FileOutputStream(file);
-            bitmap.compress(Bitmap.CompressFormat.JPEG, 100, out);
-            out.flush();
-            out.close();
-        } catch (IOException e) {
-            Log.e(TAG, "save picture error", e);
-        }
-
-        if (file.exists()) {
-            Intent data = new Intent();
-            data.setData(Uri.parse(mPath));
-//            setResult(RESULT_OK, data);
-        }
-        CameraFrag.setImage();
-        Intent intent = new Intent(getContext(), ModelActivity.class);
-        Bundle b = new Bundle();
-        b.putInt("countModel", CameraFrag.getCount() - 1);
-        b.putString("model", "null");
-        intent.putExtras(b);
-        startActivity(intent);
-    }
-
-    @Override
-    public void onCameraClose() {
-
-    }
-
-    @Override
-    public void onCameraError(Throwable th) {
-        Log.e(TAG, "camera error", th);
-        onCameraClose();
-    }
-
-    @Override
     public void onResume() {
         super.onResume();
-        mCameraView.onResume();
+//        mCameraView.onResume();
+        Log.d(TAG, "onResume");
+
+        if (!CameraManager.getInstance().isOpened() && isSurfaceCreated) {
+            CameraManager.getInstance().open(success -> {
+                if (!success && mCameraListener != null) {
+                    mCameraListener.onCameraError(new Exception("open camera failed"));
+                }
+            });
+        }
     }
 
     @Override
     public void onPause() {
         super.onPause();
-        mCameraView.onPause();
+        Log.d(TAG, "onPause");
+
+        if (CameraManager.getInstance().isOpened()) {
+            CameraManager.getInstance().close();
+        }
+//        mCameraView.onPause();
+    }
+
+
+    public interface CameraListener {
+        void onCapture(Bitmap bitmap);
+
+        void onCameraError(Throwable th);
+
+    }
+
+
+//    private void init() {
+//
+//    }
+
+    @Override
+    public void surfaceCreated(SurfaceHolder holder) {
+        Log.d(TAG, "surfaceCreated");
+        isSurfaceCreated = true;
+    }
+
+    @Override
+    public void surfaceChanged(SurfaceHolder holder, int format, int width, int height) {
+        Log.d(TAG, "surfaceChanged");
+        CameraManager.getInstance().setSurfaceHolder(holder, width, height);
+
+        if (CameraManager.getInstance().isOpened()) {
+            CameraManager.getInstance().close();
+        }
+
+        CameraManager.getInstance().open(success -> {
+            if (!success && mCameraListener != null) {
+                mCameraListener.onCameraError(new Exception("open camera failed"));
+            }
+        });
+    }
+
+    @Override
+    public void surfaceDestroyed(final SurfaceHolder holder) {
+        Log.d(TAG, "surfaceDestroyed");
+        isSurfaceCreated = false;
+        CameraManager.getInstance().setSurfaceHolder(null, 0, 0);
+    }
+
+    public void onCaptureClick() {
+        CameraManager.getInstance().takePicture(bitmap -> {
+            if (bitmap != null) {
+                mSurfaceView.setVisibility(GONE);
+                mSwitchWrapper.setVisibility(GONE);
+                btnSheetOpen.setVisibility(GONE);
+                mPictureView.setVisibility(VISIBLE);
+                mPicture = bitmap;
+                mPictureView.setImageBitmap(mPicture);
+                expand();
+            } else {
+                isExpanded = false;
+                onRetryClick();
+            }
+        });
+    }
+
+    public void onOkClick() {
+        if (mPicture != null && mCameraListener != null) {
+            mCameraListener.onCapture(mPicture);
+        }
+    }
+
+    public void onRetryClick() {
+        mPicture = null;
+        mSurfaceView.setVisibility(VISIBLE);
+        mSwitchWrapper.setVisibility(CameraManager.getInstance().hasMultiCamera() ? VISIBLE : GONE);
+        btnSheetOpen.setVisibility(VISIBLE);
+        mPictureView.setImageBitmap(null);
+        mPictureView.setVisibility(GONE);
+        fold();
+    }
+
+    public void onSwitchClick() {
+        CameraManager.getInstance().switchCamera(success -> {
+            if (!success && mCameraListener != null) {
+                mCameraListener.onCameraError(new Exception("switch camera failed"));
+            }
+        });
+    }
+
+    @Override
+    public void onClick(View v) {
+        if (v == mSwitchWrapper) {
+            Log.d(TAG, "info click");
+        }
+
+        if (v == btnCapture) {
+            if (!isExpanded) {
+                onCaptureClick();
+            } else {
+                onOkClick();
+            }
+        } else if (v == btnRetry) {
+            onRetryClick();
+        } else if (v == btnSwitchCamera) {
+            onSwitchClick();
+        }
+    }
+
+    private void expand() {
+        isExpanded = true;
+        btnCapture.setImageResource(R.drawable.ic_camera_done);
+        btnRetry.setEnabled(true);
+        btnSwitchCamera.setVisibility(GONE);
+
+        playExpandAnimation();
+    }
+
+    @TargetApi(Build.VERSION_CODES.HONEYCOMB)
+    private void playExpandAnimation() {
+        ValueAnimator scaleAnimator = ValueAnimator.ofInt(CameraUtils.dp2px(getContext(), 60), CameraUtils.dp2px(getContext(), 80));
+        scaleAnimator.setInterpolator(new LinearInterpolator());
+        scaleAnimator.setDuration(100);
+        scaleAnimator.addUpdateListener(animation -> {
+            int value = (int) animation.getAnimatedValue();
+            FrameLayout.LayoutParams captureParams = (FrameLayout.LayoutParams) btnCapture.getLayoutParams();
+            captureParams.width = value;
+            captureParams.height = value;
+            captureParams.gravity = Gravity.CENTER;
+            btnCapture.requestLayout();
+        });
+
+        ValueAnimator transAnimator = ValueAnimator.ofInt(CameraUtils.dp2px(getContext(), 80), CameraUtils.dp2px(getContext(), 280));
+        transAnimator.setInterpolator(new LinearInterpolator());
+        transAnimator.setDuration(200);
+        transAnimator.addUpdateListener(animation -> {
+            int value = (int) animation.getAnimatedValue();
+            FrameLayout.LayoutParams captureParams = (FrameLayout.LayoutParams) btnCapture.getLayoutParams();
+            captureParams.gravity = Gravity.END;
+
+            CoordinatorLayout.LayoutParams layoutParams = (CoordinatorLayout.LayoutParams) captureRetryLayout.getLayoutParams();
+            layoutParams.width = value;
+            captureRetryLayout.requestLayout();
+        });
+
+        AnimatorSet animatorSet = new AnimatorSet();
+        animatorSet.playSequentially(scaleAnimator, transAnimator);
+        animatorSet.start();
+    }
+
+    private void fold() {
+        isExpanded = false;
+        btnCapture.setImageResource(0);
+        btnRetry.setEnabled(false);
+        btnSwitchCamera.setVisibility(VISIBLE);
+
+        int length = CameraUtils.dp2px(getContext(), 60);
+        FrameLayout.LayoutParams captureParams = (FrameLayout.LayoutParams) btnCapture.getLayoutParams();
+        captureParams.width = length;
+        captureParams.height = length;
+        captureParams.gravity = Gravity.CENTER;
+
+        CoordinatorLayout.LayoutParams layoutParams = (CoordinatorLayout.LayoutParams) captureRetryLayout.getLayoutParams();
+        layoutParams.width = CameraUtils.dp2px(getContext(), 80);
+        captureRetryLayout.requestLayout();
     }
 }
