--- app/src/main/res/layout/camera_capture_layout.xml
@@ -1,51 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<merge xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content">
-
-    <ImageView
-        android:id="@+id/camera_close"
-        android:layout_width="40dp"
-        android:layout_height="40dp"
-        android:layout_gravity="center_vertical"
-        android:layout_marginLeft="50dp"
-        android:contentDescription="@null"
-        android:scaleType="centerInside"
-        android:src="@drawable/ic_camera_sw" />
-
-    <FrameLayout
-        android:id="@+id/camera_capture_retry_layout"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="center">
-
-        <ImageView
-            android:id="@+id/camera_retry"
-            android:layout_width="80dp"
-            android:layout_height="80dp"
-            android:background="@drawable/camera_btn_retry_bg"
-            android:contentDescription="@null"
-            android:scaleType="centerInside"
-            android:src="@drawable/ic_camera_retry" />
-
-        <ImageView
-            android:id="@+id/camera_capture"
-            android:layout_width="60dp"
-            android:layout_height="60dp"
-            android:layout_gravity="center"
-            android:background="@drawable/camera_btn_capture_bg"
-            android:contentDescription="@null"
-            android:scaleType="centerInside" />
-    </FrameLayout>
-
-    <ImageView
-        android:id="@+id/btn_sheet_open"
-        android:layout_width="40dp"
-        android:layout_height="40dp"
-        android:layout_gravity="center|end"
-        android:layout_marginEnd="50dp"
-        android:contentDescription="@null"
-        android:scaleType="centerInside"
-        android:src="@drawable/ic_add" />
-
-</merge>
