+++ app/src/main/res/layout/camera_fragment.xml
@@ -4,10 +4,86 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent">
 
-    <com.orego.corporation.orego.views.cameraview.CameraView
-        android:id="@+id/camera_view"
+    <SurfaceView
+        android:id="@+id/camera_surface"
         android:layout_width="match_parent"
         android:layout_height="match_parent" />
+
+
+    <FrameLayout
+        android:id="@+id/camera_switch_wrapper"
+        android:layout_width="80dp"
+        android:layout_height="80dp"
+        android:layout_gravity="end">
+
+        <ImageView
+            android:id="@+id/camera_switch"
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"
+            android:contentDescription="@null"
+            android:scaleType="centerInside"
+            android:src="@drawable/ic_info" />
+    </FrameLayout>
+
+    <ImageView
+        android:id="@+id/camera_picture_preview"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:contentDescription="@null"
+        android:scaleType="fitCenter" />
+
+    <!--<com.orego.corporation.orego.views.cameraview.CaptureLayout-->
+    <!--android:id="@+id/camera_capture_layout"-->
+    <!--android:layout_width="match_parent"-->
+    <!--android:layout_height="wrap_content"-->
+    <!--android:layout_gravity="bottom"-->
+    <!--android:paddingBottom="30dp"-->
+    <!--android:paddingTop="30dp" />-->
+
+    <ImageView
+        android:id="@+id/camera_close"
+        android:layout_width="40dp"
+        android:layout_height="40dp"
+        android:layout_gravity="bottom|start"
+        android:layout_marginLeft="50dp"
+        android:contentDescription="@null"
+        android:scaleType="centerInside"
+        android:src="@drawable/ic_camera_sw" />
+
+    <FrameLayout
+        android:id="@+id/camera_capture_retry_layout"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_gravity="bottom|center">
+
+        <ImageView
+            android:id="@+id/camera_retry"
+            android:layout_width="80dp"
+            android:layout_height="80dp"
+            android:background="@drawable/camera_btn_retry_bg"
+            android:contentDescription="@null"
+            android:scaleType="centerInside"
+            android:src="@drawable/ic_camera_retry" />
+
+        <ImageView
+            android:id="@+id/camera_capture"
+            android:layout_width="60dp"
+            android:layout_height="60dp"
+            android:layout_gravity="center"
+            android:background="@drawable/camera_btn_capture_bg"
+            android:contentDescription="@null"
+            android:scaleType="centerInside" />
+    </FrameLayout>
+
+    <ImageView
+        android:id="@+id/btn_sheet_open"
+        android:layout_width="40dp"
+        android:layout_height="40dp"
+        android:layout_gravity="bottom|end"
+        android:layout_marginEnd="50dp"
+        android:contentDescription="@null"
+        android:scaleType="centerInside"
+        android:src="@drawable/ic_add" />
 
     <TextView
         android:id="@+id/text_view_state"
