--- app/src/main/res/layout/camera_layout.xml
@@ -1,45 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<merge xmlns:android="http://schemas.android.com/apk/res/android">
-
-    <SurfaceView
-        android:id="@+id/camera_surface"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent" />
-
-    <com.orego.corporation.orego.views.cameraview.FocusView
-        android:id="@+id/camera_focus"
-        android:layout_width="120dp"
-        android:layout_height="120dp"
-        android:visibility="invisible" />
-
-    <FrameLayout
-        android:id="@+id/camera_switch_wrapper"
-        android:layout_width="80dp"
-        android:layout_height="80dp"
-        android:layout_gravity="end">
-
-        <ImageView
-            android:id="@+id/camera_switch"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:contentDescription="@null"
-            android:scaleType="centerInside"
-            android:src="@drawable/ic_info" />
-    </FrameLayout>
-
-    <ImageView
-        android:id="@+id/camera_picture_preview"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:contentDescription="@null"
-        android:scaleType="fitCenter" />
-
-    <com.orego.corporation.orego.views.cameraview.CaptureLayout
-        android:id="@+id/camera_capture_layout"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_gravity="bottom"
-        android:paddingBottom="30dp"
-        android:paddingTop="30dp" />
-
-</merge>
